refactor: Aggregate pkg
