+++ internal/pkg/stacktrace/stacktrace.go
