+++ internal/pkg/stacktrace/stacktrace_test.go
